* Enables to run single node tests * Forces copy to server flag for database add * Fixes failing tests due to 7.8.0 upgrade
