+++ .env
@@ -14,6 +14,7 @@ HAPROXY_TAG=latest
 # CircleCI does not resolve hostnames defined in docker-compose easily (i.e using the hostname field).
 # Working around this using the container name as hostnames.
 # https://www.docker.com/blog/docker-1-10/
+STARDOG_HOSTNAME_SINGLE_NODE=stardog
 STARDOG_HOSTNAME_NODE_1=stardog-1
 STARDOG_HOSTNAME_NODE_2=stardog-2
 STARDOG_HOSTNAME_STANDBY=stardog-standby
