+++ docker-compose.single-node.yml
@@ -1,0 +1,31 @@
+# Note that circleci docker executor does not allow to volume mounting, hence the volume block can't be used here.
+# https://support.circleci.com/hc/en-us/articles/360007324514-How-can-I-use-Docker-volume-mounting-on-CircleCI-
+# For this reason, we are creating specific dockerfiles, and building the images from them, copying the files in the images
+# instead of mounting them.
+
+version: "3.3"
+services:
+
+  # So another node that does not depends on zookeper.
+  sd-single-node:
+    build:
+      context: dockerfiles/
+      dockerfile: dockerfile-stardog
+      args:
+        - TAG=${STARDOG_TAG}
+        - NODE_TYPE=single
+    container_name: ${STARDOG_HOSTNAME_SINGLE_NODE}
+    entrypoint: ["/bin/bash", "-c"]
+    command: ["/var/start.sh"]
+
+  tests-single-node:
+    build:
+      context: .
+      dockerfile: dockerfiles/dockerfile-python
+    entrypoint: /bin/bash -c
+    command: ["./utils/run_test_single_node.sh"]
+    container_name: pystardog_tests_single_node
+    environment:
+      - STARDOG_ENDPOINT=${STARDOG_HOSTNAME_SINGLE_NODE}
+      - STARDOG_USER=${STARDOG_USER}
+      - STARDOG_PASS=${STARDOG_PASS}
