+++ dockerfiles/stardog.single.properties
@@ -1,0 +1,2 @@
+# Empty as we currently don't need specific properties for single node testing
+# we still need an existing property file so that the Dockerfile doesn't break
