+++ utils/run_test_single_node.sh
@@ -1,0 +1,8 @@
+#!/bin/bash -x
+
+ls -lha
+pwd
+source ./utils/wait.sh
+
+wait_for_start_single_node ${STARDOG_ENDPOINT}  5820
+pytest test/test_single_node.py --endpoint http://${STARDOG_ENDPOINT}:5820 -s
